update `@tauri-apps/api` and `@tauri-apps/plugin-shell` to match Cargo versions
In Tauri *CLI* versions >=2.8.0, the build fails if NPM and Cargo versions
are not aligned. This is not a new requirement, this merely verifies an
existing requirement. The CLI version in Nixpkgs was updated and now
detects this already-existing error.

This pulls the NPM versions up to match the existing Cargo versions
without touching any other dependencies.