Add a new commit if we just finished a txn that has an immutable head
This copies over the behavior jj-cli shows when making the currently-checked-out rev turns immutable. Also adds a test mutations::immutable_workspace_head that exercises this behavior.

In order to make the check for immutability of the new head rev work, this change also refactors various evaluation methods to take a repo reference, which call sites provide as the current operation's repo - but in-flight transactions now can provide the repo as it is currently being seen.